ggot ninja file down to 30 lines, next refactor is how to use it with write file.
